+++ concurrency/Control/Concurrent/Classy/STM.hs
@@ -15,6 +15,7 @@ module Control.Concurrent.Classy.STM
   , module Control.Concurrent.Classy.STM.TQueue
   , module Control.Concurrent.Classy.STM.TBQueue
   , module Control.Concurrent.Classy.STM.TArray
+  , module Control.Concurrent.Classy.STM.TSem
   ) where
 
 import           Control.Concurrent.Classy.STM.TArray
@@ -22,5 +23,6 @@ import           Control.Concurrent.Classy.STM.TBQueue
 import           Control.Concurrent.Classy.STM.TChan
 import           Control.Concurrent.Classy.STM.TMVar
 import           Control.Concurrent.Classy.STM.TQueue
+import           Control.Concurrent.Classy.STM.TSem
 import           Control.Concurrent.Classy.STM.TVar
 import           Control.Monad.STM.Class